requirements.txt
aiohttp==3.10.11
polars==1.8.2
pandas==1.5.3
scikit-learn==1.3.2
lightgbm==4.6.0
pyyaml==6.0.1
joblib==1.3.2
requests==2.31.0
prometheus_client==0.20.0
